% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{add_parameters}
\alias{add_parameters}
\title{Enhance Hydrofabric Network}
\usage{
add_parameters(
  gpkg = NULL,
  lyrs = c("forcing_weights", "model_attributes"),
  base_s3 = "s3://lynker-spatial/v20.1"
)
}
\arguments{
\item{gpkg}{hydrofabric id (relevant only to nextgen fabrics)}

\item{lyrs}{layers to extract. Default is "forcing_weights" and "model_attributes"}

\item{base_s3}{the base hydrofabric directory to access in Lynker's s3}
}
\value{
file path of gpkg
}
\description{
Enhance Hydrofabric Network
}
