% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qml.R
\name{append_style}
\alias{append_style}
\title{Append style to GPKG}
\usage{
append_style(
  gpkg_path,
  qml_dir = system.file("qml", package = "hydrofabric"),
  layer_names
)
}
\arguments{
\item{gpkg_path}{GPKG path}

\item{qml_dir}{Directory path}

\item{layer_names}{layer names to populate}
}
\value{
gpkg path
}
\description{
Append style to GPKG
}
