% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_arrow.R
\name{arrow_to_sf}
\alias{arrow_to_sf}
\title{Convert Arrow Table to sf}
\usage{
arrow_to_sf(tbl, metadata)
}
\arguments{
\item{tbl}{\code{data.frame} from reading an Arrow dataset}

\item{metadata}{\code{list} of validated geo metadata}
}
\value{
object of \code{sf} with CRS and geometry columns
}
\description{
Helper function to convert 'data.frame' to \code{sf}
}
\details{
Adopted from \href{https://github.com/wcjochem/sfarrow}{wcjochem/sfarrow}
}
\keyword{internal}
