% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_arrow.R
\name{create_metadata}
\alias{create_metadata}
\title{Create standardized geo metadata for Parquet files}
\usage{
create_metadata(
  df,
  hf_version = "2.2",
  license = "ODbL",
  source = "lynker-spatial"
)
}
\arguments{
\item{df}{object of class \code{sf}}

\item{hf_version}{dataset version}

\item{license}{dataset license}

\item{source}{dataset source}
}
\value{
JSON formatted list with geo-metadata
}
\description{
Create standardized geo metadata for Parquet files
}
\details{
Reference for metadata standard:
  \url{https://github.com/geopandas/geo-arrow-spec}. This is compatible with
  \code{GeoPandas} Parquet files. Adopted from \href{https://github.com/wcjochem/sfarrow}{wcjochem/sfarrow}
}
\keyword{internal}
