% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qml.R
\name{create_style_row}
\alias{create_style_row}
\title{Create style row}
\usage{
create_style_row(gpkg_path, layer_name, style_name, style_qml)
}
\arguments{
\item{gpkg_path}{GPKG path}

\item{layer_name}{layer name}

\item{style_name}{style name}

\item{style_qml}{style QML}
}
\value{
data.frame
}
\description{
Create style row
}
