% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_arrow.R
\name{encode_wkb}
\alias{encode_wkb}
\title{Encode Well Known Binary}
\usage{
encode_wkb(df)
}
\arguments{
\item{df}{\code{sf} object}
}
\value{
\code{data.frame} with binary geometry column(s)
}
\description{
Convert \code{sfc} geometry columns into a WKB binary format
}
\details{
Allows for more than one geometry column in \code{sfc} format. 
Adopted from \href{https://github.com/wcjochem/sfarrow}{wcjochem/sfarrow}
}
\keyword{internal}
