% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{get_fabric}
\alias{get_fabric}
\title{Access Hydrofabric Network}
\usage{
get_fabric(
  VPU,
  base_s3 = "s3://lynker-spatial/pre-release/",
  cache_dir = NULL,
  cache_overwrite = FALSE
)
}
\arguments{
\item{VPU}{Vector Processing Unit}

\item{base_s3}{the base hydrofabric directory to access in Lynker's s3}

\item{cache_dir}{should data be cached to a local directory? Will speed up multiple subsets in the same region}

\item{cache_overwrite}{description. Should a cached file be overwritten}
}
\value{
file path
}
\description{
Access Hydrofabric Network
}
