% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpu_fabric.R
\name{get_vpu_fabric}
\alias{get_vpu_fabric}
\title{Extract VPU GPKG from Lynker-Spatial
Data from a Parquet Dataset will be extracted for a VPU. 
Optionally this data can be written to a GPKG.}
\usage{
get_vpu_fabric(
  vpu = "01",
  type = "reference",
  version = "2.2",
  source = "s3://lynker-spatial/hydrofabric",
  outfile = NULL
)
}
\arguments{
\item{vpu}{VPU ID}

\item{type}{hydrofabric type}

\item{version}{hydrofabric version}

\item{source}{hydrofabric source (local root directory or s3 link)}

\item{outfile}{Data will be written to a file if supplied. 
Must be a gpkg file path. 
If NULL a list object of divides, flowpaths, and a network table will be returned.}
}
\description{
Extract VPU GPKG from Lynker-Spatial
Data from a Parquet Dataset will be extracted for a VPU. 
Optionally this data can be written to a GPKG.
}
