% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/zzz.R
\docType{data}
\name{hf_dm}
\alias{hf_dm}
\title{Hydrofabric Data Model}
\format{
An object of class \code{list} of length 7.

An object of class \code{list} of length 7.
}
\usage{
hf_dm

hf_dm
}
\description{
Hydrofabric Data Model

Hydrofabric Data Model
}
\keyword{datasets}
