% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{hydrofabric-package}
\alias{hydrofabric}
\alias{hydrofabric-package}
\title{hydrofabric: Easily Install and Load 'hydrofabric' tools}
\description{
The 'hydrofabric' meta package loads a set of packages that work in harmony to complete hydroscience tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NOAA-OWP/hydrofabric}
  \item Report bugs at \url{https://github.com/NOAA-OWP/hydrofabric}
}

}
\author{
\strong{Maintainer}: Mike Johnson \email{mike.johnson@noaa.gov}

Other contributors:
\itemize{
  \item NOAA [funder]
  \item USGS [contributor]
  \item Lynker [funder]
}

}
\keyword{internal}
