% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hydrofabric_packages}
\alias{hydrofabric_packages}
\title{List all packages in the hydrofabric}
\usage{
hydrofabric_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include hydrofabric in the list?}
}
\description{
List all packages in the hydrofabric
}
\examples{
hydrofabric_packages()
}
