% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_reference_feature.R
\name{input_to_reference_feature}
\alias{input_to_reference_feature}
\title{Find a Refernece Feature}
\usage{
input_to_reference_feature(net, id, comid, hl_id, poi_id, nldi_feature, xy)
}
\arguments{
\item{id}{hydrofabric id (relevant only to nextgen fabrics)}

\item{comid}{NHDPlusV2 COMID}

\item{hl_id}{hydrolocation URI (relevant only to nextgen fabrics)}

\item{poi_id}{POI identifier}

\item{nldi_feature}{list with names 'featureSource' and 'featureID' where 'featureSource' is derived from the "source" column of the response of dataRetrieval::get_nldi_sources() and the 'featureID' is a known identifier from the specified 'featureSource'.}

\item{xy}{Location given as vector of XY in CRS 4326 (longitude, latitude)}

\item{network}{table from network file default is NULL}
}
\description{
Find a Refernece Feature
}
