% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{is.url}
\alias{is.url}
\title{Is URL}
\usage{
is.url(x)
}
\arguments{
\item{x}{character string}
}
\value{
boolean
}
\description{
Identifies if a character string is a URL by searching 
for www, http, or https
}
