% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_arrow.R
\name{read_sf_dataset}
\alias{read_sf_dataset}
\title{Read Parquet Dataset}
\usage{
read_sf_dataset(dataset, find_geom = FALSE)
}
\arguments{
\item{dataset}{a \code{Dataset} object created by \code{arrow::open_dataset}
or an \code{arrow_dplyr_query}}

\item{find_geom}{logical. Only needed when returning a subset of columns.
Should all available geometry columns be selected and added to to the
dataset query without being named? Default is \code{FALSE} to require
geometry column(s) to be selected specifically.}
}
\value{
object of class \code{\link[sf]{sf}}
}
\description{
Read an Arrow multi-file dataset and create \code{sf} object
}
\details{
This function is primarily for use after opening a dataset with
  \code{arrow::open_dataset}. Users can then query the \code{arrow Dataset}
  using \code{dplyr} methods such as \code{\link[dplyr]{filter}} or
  \code{\link[dplyr]{select}}. Passing the resulting query to this function
  will parse the datasets and create an \code{sf} object. The function
  expects consistent geographic metadata to be stored with the dataset in
  order to create \code{\link[sf]{sf}} objects. 
  Adopted from \href{https://github.com/wcjochem/sfarrow}{wcjochem/sfarrow}
}
\seealso{
\code{\link[arrow]{open_dataset}}, \code{\link[sf]{st_read}}, \code{\link{st_read_parquet}}
}
