% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{read_sf_dataset_sqlite}
\alias{read_sf_dataset_sqlite}
\title{Extract spatial data from an SQLite connection}
\usage{
read_sf_dataset_sqlite(tbl)
}
\arguments{
\item{tbl}{remote temporary table}
}
\value{
an sf object (or data.frame is non spatial)
}
\description{
Extract spatial data from an SQLite connection
}
