% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_arrow.R
\name{st_write_parquet}
\alias{st_write_parquet}
\title{Write \code{sf} object to Parquet file}
\usage{
st_write_parquet(
  obj,
  dsn,
  hf_version = "2.2",
  license = "ODbL",
  source = "lynker-spatial",
  ...
)
}
\arguments{
\item{obj}{object of class \code{\link[sf]{sf}}}

\item{dsn}{data source name. A path and file name with .parquet extension}

\item{hf_version}{dataset version}

\item{license}{dataset license}

\item{source}{dataset source}

\item{...}{additional options to pass to \code{\link[arrow]{write_parquet}}}
}
\value{
\code{obj} invisibly
}
\description{
Convert a simple features spatial object from \code{sf} 
  to a Parquet file using \code{\link[arrow]{write_parquet}}. Geometry
  columns (type \code{sfc}) are converted to well-known binary (WKB) format.
}
\details{
Adopted from \href{https://github.com/wcjochem/sfarrow}{wcjochem/sfarrow}
}
\seealso{
\code{\link[arrow]{write_parquet}}
}
