% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{subset_bbox}
\alias{subset_bbox}
\title{Subset a Hydrofabric by Bounding Box}
\usage{
subset_bbox(
  gpkg,
  bbox,
  lyrs = c("divides", "nexus", "flowpaths", "network", "hydrolocations",
    "reference_flowline", "reference_catchment", "refactored_flowpaths",
    "refactored_divides"),
  outfile = NULL,
  qml_dir = system.file("qml", package = "hydrofabric")
)
}
\arguments{
\item{gpkg}{a path to a gpkg to subset}

\item{bbox}{a numeric vector of length four, with xmin, ymin, xmax and ymax values}

\item{lyrs}{layers to extract. Default is all possible in the hydrofabric GPKG data model}

\item{outfile}{file path to write to. Must have ".gpkg" extension}
}
\value{
file path (outfile) or list of features
}
\description{
Subset a Hydrofabric by Bounding Box
}
