% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{subset_reference}
\alias{subset_reference}
\title{Subset Hydrofabric Network}
\usage{
subset_reference(
  nldi_feature,
  gpkg = NULL,
  pattern = "/Volumes/Transcend/ngen/CONUS-hydrofabric/01_reference/reference_{vpu}.gpkg",
  export_gpkg = NULL
)
}
\arguments{
\item{pattern}{Pattern for distributed VPU based GPKGS}

\item{export_gpkg}{file path to write to}

\item{id}{hydrofabric id}

\item{comid}{NHDPlusV2 COMID}

\item{hl_id}{Hydrolocation URI}

\item{network}{Network Parquet file}

\item{lyrs}{layers to extract}
}
\value{
file path or list
}
\description{
Subset Hydrofabric Network
}
