% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_arrow.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate Metadata}
\usage{
validate_metadata(metadata)
}
\arguments{
\item{metadata}{list for geo metadata}
}
\value{
None. Throws an error and stops execution
}
\description{
Basic checking of key geo metadata columns
}
\details{
Adopted from \href{https://github.com/wcjochem/sfarrow}{wcjochem/sfarrow}
}
\keyword{internal}
