% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_arrow.R
\name{write_sf_dataset}
\alias{write_sf_dataset}
\title{Write Parquet Dataset}
\usage{
write_sf_dataset(
  obj,
  path,
  format = "parquet",
  partitioning = dplyr::group_vars(obj),
  hf_version = "2.2",
  license = "ODbL",
  source = "lynker-spatial",
  ...
)
}
\arguments{
\item{obj}{object of class \code{\link[sf]{sf}}}

\item{path}{string path referencing a directory for the output}

\item{format}{output file format ("parquet" or "feather")}

\item{partitioning}{character vector of columns in \code{obj} for grouping or
the \code{dplyr::group_vars}}

\item{hf_version}{dataset version}

\item{license}{dataset license}

\item{source}{dataset source}

\item{...}{additional arguments and options passed to \code{arrow::write_dataset}}
}
\value{
\code{obj} invisibly
}
\description{
Write \code{sf} object to an Arrow multi-file dataset
}
\details{
Translate an \code{sf} spatial object to \code{data.frame} with WKB
  geometry columns and then write to an \code{arrow} dataset with
  partitioning. Allows for \code{dplyr} grouped datasets (using
  \code{\link[dplyr]{group_by}}) and uses those variables to define
  partitions. Adopted from \href{https://github.com/wcjochem/sfarrow}{wcjochem/sfarrow}
}
\seealso{
\code{\link[arrow]{write_dataset}}, \code{\link{st_read_parquet}}
}
