% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{area_length_filter}
\alias{area_length_filter}
\title{Area Length Thresholds}
\usage{
area_length_filter(
  tmap,
  gpkg,
  areasqkm = NULL,
  pathlengthkm = NULL,
  ms_pathlengthkm = NULL
)
}
\arguments{
\item{tmap}{Topologic Map}

\item{gpkg}{a geopackage}

\item{areasqkm}{the maximum upstream area to subset}

\item{pathlengthkm}{the maximum upstream path length to subset}

\item{ms_pathlengthkm}{the maximum upstream mainstem path length to subset}
}
\value{
data.frame
}
\description{
Area Length Thresholds
}
