% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{subset_network}
\alias{subset_network}
\title{Subset Hydrofabric Network}
\usage{
subset_network(
  id = NULL,
  comid = NULL,
  hl_id = NULL,
  network = "data/conus_net.parquet",
  pattern = "/Volumes/Transcend/ngen/CONUS-hydrofabric/05_nextgen/nextgen_{vpu}.gpkg",
  lyrs = c("divides", "nexus", "flowpaths", "network", "hydrolocations"),
  export_gpkg = NULL
)
}
\arguments{
\item{id}{hydrofabric id}

\item{comid}{NHDPlusV2 COMID}

\item{hl_id}{Hydrolocation URI}

\item{network}{Network Parquet file}

\item{pattern}{Pattern for distributed VPU based GPKGS}

\item{lyrs}{layers to extract}

\item{export_gpkg}{file path to write to}
}
\value{
file path or list
}
\description{
Subset Hydrofabric Network
}
